eldie_chat_api/requirements.txt
Django
djangorestframework
elasticsearch
pycryptodome
openai
